SNOW-1791556: Use docstrings folder for Index methods and properties
