+++ src/snowflake/snowpark/modin/plugin/docstrings/__init__.py
@@ -11,6 +11,7 @@ from snowflake.snowpark.modin.plugin.docstrings.groupby import (
     DataFrameGroupBy,
     SeriesGroupBy,
 )
+from snowflake.snowpark.modin.plugin.docstrings.index import Index
 from snowflake.snowpark.modin.plugin.docstrings.io import *  # noqa: F401,F403
 from snowflake.snowpark.modin.plugin.docstrings.resample import Resampler
 from snowflake.snowpark.modin.plugin.docstrings.series import Series
@@ -31,4 +32,5 @@ __all__ = [
     "Series",
     "SeriesGroupBy",
     "StringMethods",
+    "Index",
 ]
